+++ packages/isaacscript-common/src/features/extraConsoleCommands/v.ts
@@ -1,5 +1,14 @@
 const v = {
   persistent: {
+    darkness: false, // 1
+    labyrinth: false, // 2
+    lost: false, // 3
+    unknown: false, // 4
+    cursed: false, // 5
+    maze: false, // 6
+    blind: false, // 7
+    giant: false, // 8
+
     disableCurses: false,
   },