+++ packages/docs/docs/main/change-log.md
@@ -28,6 +28,14 @@ This page lists the changes to the IsaacScript framework.
   - `TELEPORTER_ACTIVATION_DISTANCE`
 - Added the following custom commands:
   - `dadsNote` (to warp to the Mausoleum 2 Boss Room with Dad's Note in it)
+  - `darkness` (for permanent Curse of Darkness)
+  - `labyrinth` (for permanent Curse of the Labyrinth)
+  - `lost` (for permanent Curse of the Lost)
+  - `unknown` (for permanent Curse of the Unknown)
+  - `cursed` (for permanent Curse of the Cursed)
+  - `maze` (for permanent Curse of the Maze)
+  - `blind` (for permanent Curse of the Blind)
+  - `giant` (for permanent Curse of the Giant)
 
 ## August 18th, 2022